feat: createMaskForPhone in utilities
